+++ src/components/Cataloguing/Marc/Record/index.js
@@ -10,8 +10,9 @@ import {
   Button,
   Col,
   Icon,
-  MenuSection,
+  PaneFooter
 } from '@folio/stripes/components';
+import { FormattedMessage } from 'react-intl';
 import { AppIcon } from '@folio/stripes-core';
 import { bindActionCreators } from 'redux';
 import { union, sortBy, includes } from 'lodash';
@@ -47,6 +48,7 @@ import { formFieldValue, resolve } from '../../../../redux/helpers/Selector';
 import { TAGS, TAG_NOT_REPEATABLE } from '../../Utils/MarcConstant';
 import DataFieldForm from '../Form/DataField';
 import VariableFieldForm from '../Form/VariableField';
+
 
 class Record extends React.Component<
   Props,
@@ -393,46 +395,60 @@ class Record extends React.Component<
     return reset();
   };
 
-  actionMenu = (
-    { isEditMode } = this.state,
-    { translate } = this.props
-  ) => (
-    <Fragment>
-      <MenuSection label="Actions">
+  lastMenu = () => {
+    const { isEditMode } = this.state;
+    const { translate } = this.props;
+    return (
+      (isEditMode) ? (
         <Button
           buttonStyle="primary"
-          onClick={this.saveRecord}
           buttonClass={style.rightPosition}
+          onClick={this.deleteRecord}
           type="button"
+          disabled={false}
           marginBottom0
         >
-          {isEditMode ? (
-            <Icon icon="edit">
-              {translate({ id: 'ui-marccat.cataloging.record.edit' })}
-            </Icon>
-          ) : (
-            <Icon icon="plus-sign">
-              {translate({ id: 'ui-marccat.cataloging.record.create' })}
-            </Icon>
-          )}
+          {translate({ id: 'ui-marccat.cataloging.record.delete' })}
         </Button>
-        {isEditMode && (
-          <Button
-            buttonStyle="primary"
-            buttonClass={style.rightPosition}
-            onClick={this.deleteRecord}
-            type="button"
-            disabled={false}
-            marginBottom0
-          >
-            <Icon icon="trash">
-              {translate({ id: 'ui-marccat.cataloging.record.delete' })}
-            </Icon>
-          </Button>
-        )}
-      </MenuSection>
-    </Fragment>
-  );
+      ) : (null)
+    );
+  };
+
+  getFooter = () => {
+    /* const {
+      onCancel,
+      handleSubmit,
+      pristine,
+      submitting,
+      copy,
+    } = this.props;*/
+    const cancelButton = (
+      <Button
+        buttonStyle="default mega"
+        id="cancel-instance-edition"
+        onClick={() => this.handleClose()}
+      >
+        <FormattedMessage id="ui-marccat.button.cancel" />
+      </Button>
+    );
+    const saveButton = (
+      <Button
+        id="clickable-save-instance"
+        buttonStyle="primary mega"
+        type="submit"
+        onClick={this.saveRecord}
+      >
+        <FormattedMessage id="ui-marccat.button.saveAndClose" />
+      </Button>
+    );
+
+    return (
+      <PaneFooter
+        renderStart={cancelButton}
+        renderEnd={saveButton}
+      />
+    );
+  };
 
   render() {
     const { leaderData } = this.props;
@@ -453,9 +469,10 @@ class Record extends React.Component<
             }
             paneSub={'id. ' + bibliographicRecord.id || id}
             appIcon={<AppIcon app={C.META.ICON_TITLE} />}
-            actionMenu={this.actionMenu}
+            lastMenu={this.lastMenu()}
             dismissible
             onClose={() => this.handleClose()}
+            footer={this.getFooter()}
           >
             <Row center="xs">
               <Col xs={12} sm={6} md={8} lg={8}>