+++ src/config/constants.js
@@ -90,3 +90,8 @@ export const SEPARATOR = '-';
 export const EMPTY_STRING = '';
 export const EMPTY_SPACED_STRING = ' ';
 export const SPACED_STRING_DOUBLE_QUOTE = " "; // eslint-disable-line quotes
+
+export const SEARCH_SEGMENT = {
+  BIBLIOGRAPHIC: 'bibliographic',
+  AUTHORITY: 'authority'
+};